refactor(visual-editing-standalone): build with tsdown
