--- packages/visual-editing-standalone/.gitignore
@@ -1,1 +1,0 @@
-src/**/*.d.ts
