--- packages/visual-editing-standalone/src/create-data-attribute.ts
@@ -1,6 +1,0 @@
-export {
-  type CreateDataAttribute,
-  type CreateDataAttributeProps,
-  type WithRequired,
-  createDataAttribute,
-} from './_createDataAttribute'
