--- packages/visual-editing-standalone/src/enable-visual-editing.ts
@@ -1,10 +1,0 @@
-export type {
-  DisableVisualEditing,
-  HistoryAdapter,
-  HistoryAdapterNavigate,
-  HistoryRefresh,
-  HistoryUpdate,
-  SuspiciousStegaReport,
-  VisualEditingOptions,
-} from './_enableVisualEditing'
-export {enableVisualEditing} from './_enableVisualEditing'
