+++ packages/visual-editing-standalone/src/index.test.ts
@@ -2,10 +2,18 @@ import {createDataAttribute as createDataAttributeSource} from '@sanity/visual-e
 import {enableVisualEditing as enableVisualEditingSource} from '@sanity/visual-editing/enable-visual-editing'
 import {expect, expectTypeOf, test} from 'vitest'
 
+import pkg from '../package.json'
 import * as standalone from './index'
 
 test('exposes only the standalone runtime API', () => {
   expect(Object.keys(standalone).sort()).toEqual(['createDataAttribute', 'enableVisualEditing'])
+})
+
+test('has a single root entry point and no runtime dependencies', () => {
+  expect(Object.keys(pkg.exports).sort()).toEqual(['.', './package.json'])
+  expect(Object.keys(pkg.publishConfig.exports).sort()).toEqual(['.', './package.json'])
+  expect(pkg).not.toHaveProperty('dependencies')
+  expect(pkg).not.toHaveProperty('peerDependencies')
 })
 
 test('preserves the source package implementations and types', () => {
