+++ packages/visual-editing-standalone/src/index.ts
@@ -1,16 +1,16 @@
 export {
   type CreateDataAttribute,
   type CreateDataAttributeProps,
-  type WithRequired,
   createDataAttribute,
-} from './_createDataAttribute'
-export type {
-  DisableVisualEditing,
-  HistoryAdapter,
-  HistoryAdapterNavigate,
-  HistoryRefresh,
-  HistoryUpdate,
-  SuspiciousStegaReport,
-  VisualEditingOptions,
-} from './_enableVisualEditing'
-export {enableVisualEditing} from './_enableVisualEditing'
+  type WithRequired,
+} from './createDataAttribute'
+export {
+  type DisableVisualEditing,
+  enableVisualEditing,
+  type HistoryAdapter,
+  type HistoryAdapterNavigate,
+  type HistoryRefresh,
+  type HistoryUpdate,
+  type SuspiciousStegaReport,
+  type VisualEditingOptions,
+} from './enableVisualEditing'
