--- packages/visual-editing-standalone/tsconfig.base.json
@@ -1,9 +1,0 @@
-{
-  "extends": "@sanity/tsconfig/strictest",
-  "compilerOptions": {
-    "rootDir": ".",
-    "outDir": "dist",
-    "exactOptionalPropertyTypes": false,
-    "noUncheckedIndexedAccess": false
-  }
-}
