--- packages/visual-editing-standalone/tsconfig.build.json
@@ -1,9 +1,0 @@
-{
-  "extends": "./tsconfig.base",
-  "compilerOptions": {
-    "rootDir": "src",
-    "noCheck": true
-  },
-  "include": ["src/**/*.ts"],
-  "exclude": ["dist", "node_modules", "src/**/*.test.ts"]
-}
