+++ packages/visual-editing-standalone/tsconfig.dist.json
@@ -1,0 +1,5 @@
+{
+  "extends": "@sanity/tsconfig/strictest",
+  "include": ["src"],
+  "exclude": ["dist", "node_modules", "src/**/*.test.ts"]
+}
