+++ packages/visual-editing-standalone/tsconfig.json
@@ -1,5 +1,5 @@
 {
-  "extends": "./tsconfig.base",
-  "include": ["**/*.ts", "**/*.mjs"],
+  "extends": "@sanity/tsconfig/strictest",
+  "include": ["src", "tsdown.config.ts", "vitest.config.ts"],
   "exclude": ["dist", "node_modules"]
 }
